Fix bug where unchecked checkboxes would be reset to defaults
For whatever reason, the virtual function of the derived class was not used,
but instead the one from the base class was. I don't understand why, and I'm
certain that it worked before, because I had another device where it did indeed
still work, and it broke after flashing what should be the exact same code.

So I'm assuming something changed in the compiler. Worked around the issue by
classic copy/paste programming instead of inheritance... :(