Reduce static memory usage
Inspired by and closes #12.

Those variables don't need to be static. Looking back I think I've made them
static either because I had different plans for portal() or to be lazy and
capture them implicitly for the lambdas.

This change saves almost 300 bytes of precious RAM while the portal is not
being used.